Created a new test
I created a new test, and tested whether the "FinishedProgram" test
returned exactly 1 point. Because the method was alreay written the test
passed. I made sure the test was correct by changing the point addition
value of "FinishedProgram" from 1 point to 2 points. When the method
returns 2 points, the test fails.